requirement.txt
certifi==2019.11.28
cffi==1.13.2
cloudpickle==1.2.2
cycler==0.10.0
cytoolz==0.10.1
dask==2.9.0
decorator==4.4.1
imageio==2.6.1
interval==1.0.0
kiwisolver==1.1.0
matplotlib==3.1.1
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
natsort==7.0.1
networkx==2.4
numpy==1.17.4
olefile==0.46
pandas==1.0.1
Pillow==6.2.1
protobuf==3.11.1
pycparser==2.19
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.2
resample2d-cuda==0.0.0
scikit-image==0.15.0
scipy==1.3.2
six==1.13.0
tensorboardX==1.9
toolz==0.10.0
torch==1.1.0
torchstat==0.0.7
torchvision==0.3.0
tornado==6.0.3
